requirements.txt
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
decouple==0.0.7
idna==2.10
python-dotenv==0.15.0
requests==2.25.1
selenium==3.141.0
urllib3==1.26.3
webdriver-manager==3.3.0
